refactor(components/carousel): centralize and simplify carousel logic
- Remove redundant components `arrow-icons`, `ArrowButtons`, `StyledIconButton`, and `StyledRoot` in `src/components/carousel`.
- Replace `CarouselArrows`, `CarouselDots`, and `CarouselArrowIndex` with updated refactored logic.
- Consolidate styling into component-level `styled` elements for better maintainability.
- Refactor to use `useIsRTL` hook directly inside `CarouselArrows` and `CarouselArrowIndex` for RTL handling.
- Simplify index and navigation button logic while maintaining existing functionality.
